Fix: Refactor the repository package
