+++ src/main/java/com/project/Project/service/impl/BuildingServiceImpl.java
@@ -5,11 +5,11 @@ import com.project.Project.controller.building.dto.BuildingOptionalDto;
 import com.project.Project.domain.building.Building;
 import com.project.Project.domain.embedded.Address;
 import com.project.Project.domain.embedded.Coordinate;
-import com.project.Project.repository.RoomRepository;
 import com.project.Project.repository.building.BuildingCustomRepository;
 import com.project.Project.repository.building.BuildingRepository;
 import com.project.Project.repository.projection.building.OnlyBuildingIdAndCoord;
 import com.project.Project.repository.review.ReviewRepository;
+import com.project.Project.repository.room.RoomRepository;
 import com.project.Project.service.BuildingService;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
