+++ src/main/java/com/project/Project/service/impl/MemberServiceImpl.java
@@ -1,7 +1,7 @@
 package com.project.Project.service.impl;
 
 import com.project.Project.domain.Member;
-import com.project.Project.repository.MemberRepository;
+import com.project.Project.repository.member.MemberRepository;
 import com.project.Project.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
