+++ src/main/java/com/project/Project/service/impl/RoomServiceImpl.java
@@ -2,7 +2,7 @@ package com.project.Project.service.impl;
 
 import com.project.Project.domain.building.Building;
 import com.project.Project.domain.room.Room;
-import com.project.Project.repository.RoomRepository;
+import com.project.Project.repository.room.RoomRepository;
 import com.project.Project.service.RoomService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
