+++ src/test/java/com/project/Project/controller/validation/ReviewRequestValidationTest.java
@@ -6,7 +6,7 @@
 //import com.project.Project.controller.review.controller.ReviewRestController;
 //import com.project.Project.repository.building.BuildingRepository;
 //import com.project.Project.repository.review.ReviewRepository;
-//import com.project.Project.repository.RoomRepository;
+//import com.project.Project.repository.room.RoomRepository;
 //import com.project.Project.service.BuildingService;
 //import com.project.Project.service.ReviewService;
 //import com.project.Project.service.RoomService;
